requirements.txt
streamlit
pandas
matplotlib
python-dotenv
autogen
autogen-agentchat
autogen-ext[openai]